refactor(nvim): `globals -> core`
