+++ dotfiles/nvim/lua/ik1614/core/globals.lua
